Some core modules moved to `model` subfolder
