+++ lib/folding/foldings-manager.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import { Folding, FoldingStatus } from './folding';
+import { Folding, FoldingStatus } from './model/folding';
 import { HighlightingManager } from './highlighting-manager';
 
 export const FoldingsManagerConfigCheck = {
