+++ lib/folding/highlighting-manager.js
@@ -1,7 +1,6 @@
 'use babel';
 
 import { CompositeDisposable } from 'atom';
-import { FoldingMarker } from './folding-marker';
 import { MultiMap } from '../utils/multimap';
 import logger from '../utils/logger';
