+++ lib/folding/model/folding-marker.js
@@ -1,6 +1,5 @@
 'use babel';
 
-import { Range } from 'atom';
 import { MarkerEvent } from './marker-event';
 
 export class FoldingMarker {
