+++ lib/folding/model/folding-preview.js
