+++ lib/folding/model/folding.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import logger from '../utils/logger.js';
+import logger from '../../utils/logger.js';
 import { PseudoFolding, PseudoFoldingCheck } from './pseudo-folding.js';
 import FoldingPreview from './folding-preview.js';
 import { FoldingMarker } from './folding-marker';
