+++ lib/folding/model/marker-event.js
