+++ lib/folding/model/pseudo-folding.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import { TextEditorUtils as TEUtils } from '../utils/text-editor-utils.js';
+import { TextEditorUtils as TEUtils } from '../../utils/text-editor-utils.js';
 
 export const PseudoFoldingCheck = {
     POSSIBLE: 0,
